requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
SQLAlchemy==2.0.44
asyncpg==0.30.0
alembic==1.17.1
python-dotenv==1.2.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.12.4
